+++ src/features/optional/quality/showDreamCatcherItem/v.ts
@@ -1,4 +1,4 @@
-import { saveDataManager } from "isaacscript-common";
+import { PlayerHealth, saveDataManager } from "isaacscript-common";
 import { config } from "../../../../modConfigMenu";
 import { DreamCatcherWarpState } from "../../../../types/DreamCatcherWarpState";
 
@@ -8,11 +8,15 @@ const v = {
     warpRoomGridIndexes: [] as int[],
     displayFlagsMap: new Map<int, int>(),
     cardReadingPortalDescriptions: [] as Array<[int, Vector]>,
+    health: null as PlayerHealth | null,
 
     collectibles: [] as CollectibleType[],
 
     /** Bosses are stored as an array of: [entityType, variant] */
     bosses: [] as Array<[int, int]>,
+
+    /** Used so that the fast-travel feature can communicate with this feature. */
+    arrivedOnNewFloor: false,
   },
 };
 export default v;
@@ -27,4 +31,8 @@ function featureEnabled() {
 
 export function isDreamCatcherWarping(): boolean {
   return v.level.warpState === DreamCatcherWarpState.WARPING;
+}
+
+export function setDreamCatcherArrivedOnNewFloor(): void {
+  v.level.arrivedOnNewFloor = true;
 }