+++ docs/history.md
@@ -22,6 +22,8 @@
 - Fixed the bug where moving diagonally into a combined duality door would not grant the right room under certain conditions. (Thanks Shiny)
 - Fixed the bug where Lost Soul was incorrectly revived after a seeded death revival. (Thanks Hispa)
 - Fixed the bug where using Mega Blast would make you move incorrectly when interacting with a trapdoor or beam of light.
+- Fixed the bug where the minimap would reveal extra information during floor transitions in the player had Dream Catcher.
+- Fixed the bug where the player's health could be set to the wrong amount when going to a new floor and having Dream Catcher while playing on a set seed. (Thanks Gamonymous)
 
 ### *v0.75.0* - February 19th, 2022